Move 'ExpandIPPerms' (and tests).
