+++ internal/service/ec2/flex.go
@@ -1,12 +1,8 @@
 package ec2
 
 import (
-	"fmt"
-	"strings"
-
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 // Flattens security group identifiers into a []string, where the elements returned are the GroupIDs
@@ -28,116 +24,6 @@ type GroupIdentifier struct {
 	GroupName *string
 
 	Description *string
-}
-
-// Takes the result of flatmap.Expand for an array of ingress/egress security
-// group rules and returns EC2 API compatible objects. This function will error
-// if it finds invalid permissions input, namely a protocol of "-1" with either
-// to_port or from_port set to a non-zero value.
-func ExpandIPPerms(
-	group *ec2.SecurityGroup, configured []interface{}) ([]*ec2.IpPermission, error) {
-	vpc := aws.StringValue(group.VpcId) != ""
-
-	perms := make([]*ec2.IpPermission, len(configured))
-	for i, mRaw := range configured {
-		var perm ec2.IpPermission
-		m := mRaw.(map[string]interface{})
-
-		perm.FromPort = aws.Int64(int64(m["from_port"].(int)))
-		perm.ToPort = aws.Int64(int64(m["to_port"].(int)))
-		perm.IpProtocol = aws.String(m["protocol"].(string))
-
-		// When protocol is "-1", AWS won't store any ports for the
-		// rule, but also won't error if the user specifies ports other
-		// than '0'. Force the user to make a deliberate '0' port
-		// choice when specifying a "-1" protocol, and tell them about
-		// AWS's behavior in the error message.
-		if aws.StringValue(perm.IpProtocol) == "-1" && (aws.Int64Value(perm.FromPort) != 0 || aws.Int64Value(perm.ToPort) != 0) {
-			return nil, fmt.Errorf(
-				"from_port (%d) and to_port (%d) must both be 0 to use the 'ALL' \"-1\" protocol!",
-				*perm.FromPort, *perm.ToPort)
-		}
-
-		var groups []string
-		if raw, ok := m["security_groups"]; ok {
-			list := raw.(*schema.Set).List()
-			for _, v := range list {
-				groups = append(groups, v.(string))
-			}
-		}
-		if v, ok := m["self"]; ok && v.(bool) {
-			if vpc {
-				groups = append(groups, *group.GroupId)
-			} else {
-				groups = append(groups, *group.GroupName)
-			}
-		}
-
-		if len(groups) > 0 {
-			perm.UserIdGroupPairs = make([]*ec2.UserIdGroupPair, len(groups))
-			for i, name := range groups {
-				ownerId, id := "", name
-				if items := strings.Split(id, "/"); len(items) > 1 {
-					ownerId, id = items[0], items[1]
-				}
-
-				perm.UserIdGroupPairs[i] = &ec2.UserIdGroupPair{
-					GroupId: aws.String(id),
-				}
-
-				if ownerId != "" {
-					perm.UserIdGroupPairs[i].UserId = aws.String(ownerId)
-				}
-
-				if !vpc {
-					perm.UserIdGroupPairs[i].GroupId = nil
-					perm.UserIdGroupPairs[i].GroupName = aws.String(id)
-				}
-			}
-		}
-
-		if raw, ok := m["cidr_blocks"]; ok {
-			list := raw.([]interface{})
-			for _, v := range list {
-				perm.IpRanges = append(perm.IpRanges, &ec2.IpRange{CidrIp: aws.String(v.(string))})
-			}
-		}
-		if raw, ok := m["ipv6_cidr_blocks"]; ok {
-			list := raw.([]interface{})
-			for _, v := range list {
-				perm.Ipv6Ranges = append(perm.Ipv6Ranges, &ec2.Ipv6Range{CidrIpv6: aws.String(v.(string))})
-			}
-		}
-
-		if raw, ok := m["prefix_list_ids"]; ok {
-			list := raw.([]interface{})
-			for _, v := range list {
-				perm.PrefixListIds = append(perm.PrefixListIds, &ec2.PrefixListId{PrefixListId: aws.String(v.(string))})
-			}
-		}
-
-		if raw, ok := m["description"]; ok {
-			description := raw.(string)
-			if description != "" {
-				for _, v := range perm.IpRanges {
-					v.Description = aws.String(description)
-				}
-				for _, v := range perm.Ipv6Ranges {
-					v.Description = aws.String(description)
-				}
-				for _, v := range perm.PrefixListIds {
-					v.Description = aws.String(description)
-				}
-				for _, v := range perm.UserIdGroupPairs {
-					v.Description = aws.String(description)
-				}
-			}
-		}
-
-		perms[i] = &perm
-	}
-
-	return perms, nil
 }
 
 // Flattens an array of UserSecurityGroups into a []*GroupIdentifier
