+++ internal/service/ec2/vpc_security_group_test.go
@@ -478,6 +478,311 @@ func TestSecurityGroupIPPermGather(t *testing.T) {
 	}
 }
 
+func TestExpandIPPerms(t *testing.T) {
+	hash := schema.HashString
+
+	expanded := []interface{}{
+		map[string]interface{}{
+			"protocol":    "icmp",
+			"from_port":   1,
+			"to_port":     -1,
+			"cidr_blocks": []interface{}{"0.0.0.0/0"},
+			"security_groups": schema.NewSet(hash, []interface{}{
+				"sg-11111",
+				"foo/sg-22222",
+			}),
+			"description": "desc",
+		},
+		map[string]interface{}{
+			"protocol":  "icmp",
+			"from_port": 1,
+			"to_port":   -1,
+			"self":      true,
+		},
+	}
+	group := &ec2.SecurityGroup{
+		GroupId: aws.String("foo"),
+		VpcId:   aws.String("bar"),
+	}
+	perms, err := tfec2.ExpandIPPerms(group, expanded)
+	if err != nil {
+		t.Fatalf("error expanding perms: %v", err)
+	}
+
+	expected := []ec2.IpPermission{
+		{
+			IpProtocol: aws.String("icmp"),
+			FromPort:   aws.Int64(1),
+			ToPort:     aws.Int64(int64(-1)),
+			IpRanges: []*ec2.IpRange{
+				{
+					CidrIp:      aws.String("0.0.0.0/0"),
+					Description: aws.String("desc"),
+				},
+			},
+			UserIdGroupPairs: []*ec2.UserIdGroupPair{
+				{
+					UserId:      aws.String("foo"),
+					GroupId:     aws.String("sg-22222"),
+					Description: aws.String("desc"),
+				},
+				{
+					GroupId:     aws.String("sg-11111"),
+					Description: aws.String("desc"),
+				},
+			},
+		},
+		{
+			IpProtocol: aws.String("icmp"),
+			FromPort:   aws.Int64(1),
+			ToPort:     aws.Int64(int64(-1)),
+			UserIdGroupPairs: []*ec2.UserIdGroupPair{
+				{
+					GroupId: aws.String("foo"),
+				},
+			},
+		},
+	}
+
+	exp := expected[0]
+	perm := perms[0]
+
+	if aws.Int64Value(exp.FromPort) != aws.Int64Value(perm.FromPort) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.Int64Value(perm.FromPort),
+			aws.Int64Value(exp.FromPort))
+	}
+
+	if aws.StringValue(exp.IpRanges[0].CidrIp) != aws.StringValue(perm.IpRanges[0].CidrIp) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.IpRanges[0].CidrIp),
+			aws.StringValue(exp.IpRanges[0].CidrIp))
+	}
+
+	if aws.StringValue(exp.UserIdGroupPairs[0].UserId) != aws.StringValue(perm.UserIdGroupPairs[0].UserId) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[0].UserId),
+			aws.StringValue(exp.UserIdGroupPairs[0].UserId))
+	}
+
+	if aws.StringValue(exp.UserIdGroupPairs[0].GroupId) != aws.StringValue(perm.UserIdGroupPairs[0].GroupId) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[0].GroupId),
+			aws.StringValue(exp.UserIdGroupPairs[0].GroupId))
+	}
+
+	if aws.StringValue(exp.UserIdGroupPairs[1].GroupId) != aws.StringValue(perm.UserIdGroupPairs[1].GroupId) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[1].GroupId),
+			aws.StringValue(exp.UserIdGroupPairs[1].GroupId))
+	}
+
+	exp = expected[1]
+	perm = perms[1]
+
+	if aws.StringValue(exp.UserIdGroupPairs[0].GroupId) != aws.StringValue(perm.UserIdGroupPairs[0].GroupId) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[0].GroupId),
+			aws.StringValue(exp.UserIdGroupPairs[0].GroupId))
+	}
+}
+
+func TestExpandIPPerms_NegOneProtocol(t *testing.T) {
+	hash := schema.HashString
+
+	expanded := []interface{}{
+		map[string]interface{}{
+			"protocol":    "-1",
+			"from_port":   0,
+			"to_port":     0,
+			"cidr_blocks": []interface{}{"0.0.0.0/0"},
+			"security_groups": schema.NewSet(hash, []interface{}{
+				"sg-11111",
+				"foo/sg-22222",
+			}),
+		},
+	}
+	group := &ec2.SecurityGroup{
+		GroupId: aws.String("foo"),
+		VpcId:   aws.String("bar"),
+	}
+
+	perms, err := tfec2.ExpandIPPerms(group, expanded)
+	if err != nil {
+		t.Fatalf("error expanding perms: %v", err)
+	}
+
+	expected := []ec2.IpPermission{
+		{
+			IpProtocol: aws.String("-1"),
+			FromPort:   aws.Int64(0),
+			ToPort:     aws.Int64(0),
+			IpRanges:   []*ec2.IpRange{{CidrIp: aws.String("0.0.0.0/0")}},
+			UserIdGroupPairs: []*ec2.UserIdGroupPair{
+				{
+					UserId:  aws.String("foo"),
+					GroupId: aws.String("sg-22222"),
+				},
+				{
+					GroupId: aws.String("sg-11111"),
+				},
+			},
+		},
+	}
+
+	exp := expected[0]
+	perm := perms[0]
+
+	if aws.Int64Value(exp.FromPort) != aws.Int64Value(perm.FromPort) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.Int64Value(perm.FromPort),
+			aws.Int64Value(exp.FromPort))
+	}
+
+	if aws.StringValue(exp.IpRanges[0].CidrIp) != aws.StringValue(perm.IpRanges[0].CidrIp) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.IpRanges[0].CidrIp),
+			aws.StringValue(exp.IpRanges[0].CidrIp))
+	}
+
+	if aws.StringValue(exp.UserIdGroupPairs[0].UserId) != aws.StringValue(perm.UserIdGroupPairs[0].UserId) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[0].UserId),
+			aws.StringValue(exp.UserIdGroupPairs[0].UserId))
+	}
+
+	// Now test the error case. This *should* error when either from_port
+	// or to_port is not zero, but protocol is "-1".
+	errorCase := []interface{}{
+		map[string]interface{}{
+			"protocol":    "-1",
+			"from_port":   0,
+			"to_port":     65535,
+			"cidr_blocks": []interface{}{"0.0.0.0/0"},
+			"security_groups": schema.NewSet(hash, []interface{}{
+				"sg-11111",
+				"foo/sg-22222",
+			}),
+		},
+	}
+	securityGroups := &ec2.SecurityGroup{
+		GroupId: aws.String("foo"),
+		VpcId:   aws.String("bar"),
+	}
+
+	_, expandErr := tfec2.ExpandIPPerms(securityGroups, errorCase)
+	if expandErr == nil {
+		t.Fatal("ExpandIPPerms should have errored!")
+	}
+}
+
+func TestExpandIPPerms_nonVPC(t *testing.T) {
+	hash := schema.HashString
+
+	expanded := []interface{}{
+		map[string]interface{}{
+			"protocol":    "icmp",
+			"from_port":   1,
+			"to_port":     -1,
+			"cidr_blocks": []interface{}{"0.0.0.0/0"},
+			"security_groups": schema.NewSet(hash, []interface{}{
+				"sg-11111",
+				"foo/sg-22222",
+			}),
+		},
+		map[string]interface{}{
+			"protocol":  "icmp",
+			"from_port": 1,
+			"to_port":   -1,
+			"self":      true,
+		},
+	}
+	group := &ec2.SecurityGroup{
+		GroupName: aws.String("foo"),
+	}
+	perms, err := tfec2.ExpandIPPerms(group, expanded)
+	if err != nil {
+		t.Fatalf("error expanding perms: %v", err)
+	}
+
+	expected := []ec2.IpPermission{
+		{
+			IpProtocol: aws.String("icmp"),
+			FromPort:   aws.Int64(1),
+			ToPort:     aws.Int64(int64(-1)),
+			IpRanges:   []*ec2.IpRange{{CidrIp: aws.String("0.0.0.0/0")}},
+			UserIdGroupPairs: []*ec2.UserIdGroupPair{
+				{
+					GroupName: aws.String("sg-22222"),
+				},
+				{
+					GroupName: aws.String("sg-11111"),
+				},
+			},
+		},
+		{
+			IpProtocol: aws.String("icmp"),
+			FromPort:   aws.Int64(1),
+			ToPort:     aws.Int64(int64(-1)),
+			UserIdGroupPairs: []*ec2.UserIdGroupPair{
+				{
+					GroupName: aws.String("foo"),
+				},
+			},
+		},
+	}
+
+	exp := expected[0]
+	perm := perms[0]
+
+	if aws.Int64Value(exp.FromPort) != aws.Int64Value(perm.FromPort) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.Int64Value(perm.FromPort),
+			aws.Int64Value(exp.FromPort))
+	}
+
+	if aws.StringValue(exp.IpRanges[0].CidrIp) != aws.StringValue(perm.IpRanges[0].CidrIp) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.IpRanges[0].CidrIp),
+			aws.StringValue(exp.IpRanges[0].CidrIp))
+	}
+
+	if aws.StringValue(exp.UserIdGroupPairs[0].GroupName) != aws.StringValue(perm.UserIdGroupPairs[0].GroupName) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[0].GroupName),
+			aws.StringValue(exp.UserIdGroupPairs[0].GroupName))
+	}
+
+	if aws.StringValue(exp.UserIdGroupPairs[1].GroupName) != aws.StringValue(perm.UserIdGroupPairs[1].GroupName) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[1].GroupName),
+			aws.StringValue(exp.UserIdGroupPairs[1].GroupName))
+	}
+
+	exp = expected[1]
+	perm = perms[1]
+
+	if aws.StringValue(exp.UserIdGroupPairs[0].GroupName) != aws.StringValue(perm.UserIdGroupPairs[0].GroupName) {
+		t.Fatalf(
+			"Got:\n\n%#v\n\nExpected:\n\n%#v\n",
+			aws.StringValue(perm.UserIdGroupPairs[0].GroupName),
+			aws.StringValue(exp.UserIdGroupPairs[0].GroupName))
+	}
+}
+
 func TestAccVPCSecurityGroup_basic(t *testing.T) {
 	var group ec2.SecurityGroup
 	resourceName := "aws_security_group.test"
